[FEATURE] Landing Page - Move to Questions Page when i click button - Add the main image
